New Auto Updater

+++ SR_Plugin_Loader/SR_PluginLoader/Plugins/Plugin.cs
@@ -515,7 +515,7 @@ namespace SR_PluginLoader
             if (this.data.UPDATE_METHOD == null) throw new ArgumentNullException(String.Format("{0} Plugin has no UPDATE_METHOD specified!", this));
             
             IEnumerator iter = this.Updater.Download(this.data.UPDATE_METHOD.URL, this.file, null,
-               (int current, int total) =>
+               (float current, float total) =>
                {//Download progress
                     float f = (float)current / (float)total;
